requirements.txt
# Core runtime
pymupdf==1.26.3
ImageHash==4.3.2
Pillow==10.4.0
numpy==1.26.4
pandas==2.3.1
openai==1.98.0

# Optional: if you need to parse Excel/CSV in future extensions
openpyxl==3.1.5
xlrd==2.0.2

# API/UI backend
fastapi==0.115.4
uvicorn==0.32.0
python-multipart==0.0.12
